requirements.txt
Django==1.5.1
Pillow==2.0.0
South==0.7.6
argparse==1.2.1
dateutils==0.6.6
distribute==0.6.36
django-allauth==0.10.1
django-appconf==0.6
django-compressor==1.3
httplib2==0.8
oauth2==1.5.211
oauthlib==0.4.0
python-dateutil==2.1
python-openid==2.2.5
pytz==2013b
requests==1.2.0
requests-oauthlib==0.3.0
six==1.3.0
wsgiref==0.1.2
